feat: pss-imp-wrk-03-runners-003 - Prove the common Runner contract and inert runtime-assembly integration
